feat:css changes in auto endorsement hover image
